Fix possibility to extend transaction by unauthorized signer
It was discovered that if a DID-authorized signer proposed a Cryptid
transaction it was possible to extend that transaction with an
unauthorized signer.

The bug was fixed by only enabling allow_unauthorized on propose
whereas subsequent operations (extend, execute, close) can only be
executed by a DID authority or by the initial unauthorized signer.

PR: https://github.com/identity-com/cryptid/pull/221